Change components names with home page
